feat: Add worker shims (#48)
I think I've got the shims working in a reasonable way now that passes
CI (and the examples all run).

It also means that all workers that use the shims will report their
errors automatically.

I think there are a couple of issues with it though:
It assumes json serialisation (I think this could be made configurable)
The shims are part of the main tierkreis dependency (maybe it could be
factored out into a tierkreis-worker package?)
"glob" style inputs and outputs rely on special behaviour if there is
exactly one input whose type is Iterator or an Iterator output which is
slightly magical

Otherwise I think the PR is now ready for review