feat: Allow multiple acc_ports for LOOP (#80)
Related to https://github.com/CQCL/tierkreis/issues/61

Remove acc_port from the definition of loop node and thread all of the
outputs of the loop body to the next iteration and ultimately the output
of the loop node.

* Modify the tests and examples to use new definition.
* Create test. A graph with a loop body that has multiple outputs.
* When walking a loop node link all of the outputs into the correct
place. (All inputs to the loop will remain accessible but not modified.)