feat: implement Partial (#93)
Simple implementation of partial graphs by inserting `Loc` data into the
`GraphData`. In future, if we need more performance we can avoid the
repeated deserialisation of the GraphData by passing the fixed inputs
through in the NodeRunData at the cost of higher code complexity. The
diff looks worse than it is because I had to extract out some variables
to avoid Python circular imports.