chore: create class portmapping (#137)
* Define a decorator `@portmapping` to pick out the `NamedTuple`s whose
attributes correspond to different outputs in the graph. Note that we
don't yet allow `NamedTuple` as a `PType` but we now have a way of
adding it in a less confusing way.
* Add the `@portmapping` to workers defined in this repo.
* Code gen: fix an order of attributes and generics in classes to avoid
unnecessary code churn.
* Change `is_pnamed_model` to `is_portmapping` because all calls to
`is_pnamed_model` were for the purposes of checking if the object was a
port mapping.