docs: graphbuilder using builtins (#146)
Initial entry in series of doc pages that form a slow introduction to
graph building and running.

* Add workflow to build and deploy docs to [GitHub
Pages](https://cqcl.github.io/tierkreis/).
* Add `myst_nb` to extensions. (`myst_nb` includes `myst_parser`.)
* Add first entry in a series of 'step by step' tutorials that creates a
simple graph with no inputs and only built-in workers.
* Re-export some classes and types that are buried a bit deep.
* Add a convenience method `read_outputs` that takes care of singletons
and calls `ptype_from_bytes`.
* Configure sphinx to fail on errors in notebooks. Should be okay for
now but we might need to revisit if we write docs where we deliberately
want a failure. (Or just handle then print the exception.)