docs: add tutorials for eval and loop (#147)
Add tutorials for graph inputs, eval and loop. (It is getting harder to
give good examples without documenting the Python worker stub generation
process so I think I'll do that next...)

* (Fix variance of `Inputs` in `GraphBuilder`, which was allowing
certain undesired subtypes to be able to be passed to `eval`. Cascading
fixes to `fold.py` and the hamiltonian graph.)
* Pass through Python's randint to builtins to enable example.
* Allow single wire inputs to run_graph.
* Consolidate various re-exports to `tierkreis.models` as a generic
place for end-user access to model/data/types stuff.