chore: insist MAP nodes have integer idx (#153)
Some clean up of internals. All the rest of the node types had integer
indices and the typed graph builder only produces integers for map nodes
anyway.

https://github.com/CQCL/tierkreis/issues/149

* Update definition of `NodeStep` and `Loc.steps` to only allow integer
values for the index of a map node.
* Create `outputs_iter` function that returns an list of nodes, which is
intended to be used in the situation where we cannot deduce their number
before runtime. Use this function in place of
`storage.read_output_ports` when dealing with map nodes.
* Restrict the `PyNode.id` to be an integer again. This should be a
non-breaking change for the front-end.