feat: use idl to generate stubs (#170)
Generate worker stub files from a simple IDL based on
[TypeSpec](https://typespec.io/docs/). Initially intended as a cheap way
to support typed graph building for all non-Python workers. In the
medium-term, we might also choose to generate interface definitions in
various languages from the IDL.

* Simple test that a namespace generated by the Python worker library
generates the same stubs as the new IDL. Generics and nested models
included.
* (Use new Python generics syntax. Aligns more closely with a recursive
analysis of the types.)
* Improve the granularity of the models in `namespace.py` to get them
further away from Python `NamedTuple`s and nearer to an abstract
description of the structures we are generating. Move lookup of types
from `Namespace` into the parent class `Worker` so that `Namespace` need
not contain arbitrary `NamedTuple`s.
* Add elementary parser and combinators. Tried half a dozen existing
libraries but in the end none were sufficiently type checked. (Also the
amount of code may not be worth a dependency anyway.)
* Create specific parser combinators for parsing our TypeSpec-like IDL.
We use a decorator `@portmapping` on the TypeSpec 'model' keyword to
distinguish portmappings from structs.