example: shell script worker (#176)
Examples of both ways of adding a shell script to a graph. One that
requires a `tierkreis.sh` to be included first and another one for
simple scripts only using stdin/out. Will wait for the new
command/executor business to be finalised before writing docs.

* Add `sign` and `verify` methods to the example `auth_worker`.
* Create `openssl_worker` from a shell script with simple test.
* Create `signing_graph` example, which combines the `UvExecutor` and
the `ShellExecutor` using `MultipleExecutor`. It runs the
`openssl_worker` and passes the output key pair into the `auth_worker`
for signing and verification. Add the example to the `justfile`.
* Improve the `ShellExecutor` to look for `worker_name/main.sh` in
addition to a top-level `worker_name` script. This allows us to write
workers in the same shape as the uv workers.
* Create `StdInOut` executor that takes care of piping files to
stdin/out and touching the _done file.