chore: include more justfile commands in ci (#192)
Create a reusable workflow to setup our project environment and run a
command. Use this for the existing checks and add two more checks.

* Check examples do not error
* Check docs do not error
* Also trigger a docs deploy on release so we don't have to do this
manually.
* Remove quantinuum-sphinx because it is failing and we don't use it.

EDIT: GitHub is confused when I try to modify/delete `ci.yml` so I'm
going to try this in two separate PRs... the next will remove the old
`ci.yml`.