Nexus worker improvements, graphs and docs (#207)
EDIT: have set this to auto-merge even though I haven't written the docs
part. It should be useful to have in for the pytket-worker branch as
similar issue have arisen there.

* Move stubs from example_workers to tierkreis core package. (Just use
worker name.)
* Create Nexus polling example that submits a circuit and polls for
results. Include the graph in the core tierkreis library so others can
easily access.
* Work around UnionType problem by relying more on `format_generic_type`
and de-emphasising `format_ptype` in codegen.
* (Make the stubs tests output a file so that we can check it lints and
that it hasn't unexpectedly changed its output.)
* Add builtin sleep node for quick solution to rate limiting.
* (Allow GraphBuilder in `run_graph`.)
* Quick solution to coercing union types; revisit if performs badly with
large unions.
* Strip `Annotation` from types and just use underlying type. Required
to deal with BackendConfig.
* Re-work Nexus worker to more closely match the corresponding methods
in qnexus.