chore: include frontend assets in PyPI package (#215)
This one might take a bit of pushing through CI with the various release
processes. Tested locally by importing tierkreis_visualization
(non-editable!) into a new project and seeing a not found error. Then
with the change to `pyproject.toml` and a re-install the dist is
included in the package and the visualiser loads.