feat: default arguments python workers (#239)
Support default values for arguments in worker tasks and optional values
in models used to build graphs.

* Add example that show use of default values in worker tasks and
optionals work in graph builder objects.
* Remove code and tests for partial evaluation, which wasn't accessible
in the typed graph builder. Add issue to create a new node type for
partial.
* Handle `None` as a `TNamedModel` attribute, which will be interpreted
by worker tasks as 'argument is not present'. Send warnings when an
argument is not present rather than failing. If the graph type checks
then it should complete.
* Use `inspect.Signature` to detect which task arguments have defaults.
Use this information to avoid raising an exception when no value is
found for an argument with a default value. Update codegen to make an
argument optional in the graph builder if it has a default value.
* Update `aer_worker` to use default values. Notice that `pytket_worker`
already had specified some default values, which now get picked up.