chore: simplify path based storage  (#217)
This PR is adapted from the stale branch for adding an S3 storage.
Essentially we simplify the creation of all storages based on paths,
which we expect to be the vast majority. Other ones, like
GraphDataStorage, can override more specific methods.

* Create simpler storage interface based on methods like `delete`,
`exists`, `read`, `write`, etc... Previous interface methods are
combinations of these. As a result we now usually don't need to add a
method to the interface every time we want to add additional
functionality to the storage. We just have to write down how to do this
using the new interface.
* Rewrite existing storage implementations to use the new interface.
Filestorage and in-memory storage are much simpler. GraphDataStorage is
slightly longer.